requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
convert==0.1.2
cssselect2==0.7.0
Django==5.0.2
django-debug-toolbar==4.3.0
fonttools==4.49.0
html5lib==1.1
idna==3.6
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
urllib3==2.2.1
weasyprint==61.0
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.2.3

djangorestframework~=3.15.2